+++ api/crates/perception_infra/src/lib.rs
@@ -6,6 +6,7 @@ mod transient_annotation_repository;
 mod transient_dataset_repository;
 mod transient_dataset_version_repository;
 mod transient_sample_repository;
+mod transient_training_job_queue;
 mod transient_training_job_repository;
 
 pub use local_sample_storage::LocalSampleStorage;
@@ -13,6 +14,7 @@ pub use transient_annotation_repository::TransientAnnotationRepository;
 pub use transient_dataset_repository::TransientDatasetRepository;
 pub use transient_dataset_version_repository::TransientDatasetVersionRepository;
 pub use transient_sample_repository::TransientSampleRepository;
+pub use transient_training_job_queue::TransientTrainingJobQueue;
 pub use transient_training_job_repository::TransientTrainingJobRepository;
 
 pub const CRATE_NAME: &str = "perception_infra";